Move to 2.1.2 vertx, fix tests to allow send and publish on template compile, compile on-the-fly and from sharedData in multi verticel environment.

+++ src/test/resources/integration_tests/javascript/basicIntegrationTest.js
@@ -84,7 +84,10 @@ var script = this;
 
 function setup(){
 	console.log("Cleaning compilation folder compiledbasic/");
-	fs.deleteSync("compiledbasic/",true);
+	var result = fs.existsSync("compiledbasic/");
+	if(result){
+		fs.deleteSync("compiledbasic/",true);
+	}
 	console.log("Cleaned");
 }
